backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
passlib[bcrypt] 
pydantic
sqlalchemy
pymysql
pytest
 pytest-asyncio
